* Re-did the entire layout, added Id3 Tags for Meta Data, and added a genre dictionary.

+++ MauiMediaPlayer/MauiProgram.cs
@@ -46,37 +46,39 @@ namespace MauiMediaPlayer
 
 
             // Add Stuff Here Later // - cjm 
-            Debug.WriteLine("\n=== ======================================== ===\n");
-            Debug.WriteLine("*** WARNING: This is a debug build.  The database will be deleted and recreated.  Change this later. (MauiProgram.cs)");
-            var _dbContext = new PlaylistContext();
-            _dbContext.Database.EnsureDeleted();
-            _dbContext.Database.EnsureCreated();
-            _dbContext.Playlists.Add(new Playlist { Name = "Default", Description = "The Default Play List" });
-            _dbContext.Playlists.Add(new Playlist { Name = "Test", Description = "The Test Play List" });
-            _dbContext.Playlists.Add(new Playlist { Name = "Test2", Description = "The Test2 Play List" });
-            _dbContext.Playlists.Add(new Playlist { Name = "Test3", Description = "The Test3 Play List" });
-
-            _dbContext.Songs.Add(new Song
+            if (true)
             {
-                Title = "02 - Test Song - With a Very Long Name",
-                Artist = "Test Artist - With a Very Long Name",
-                Album = "Test Album  - With a Very Long Name",
-                Genre = "Test Genre - With a Very Long Name - And a Whole Bunch More Gibberish",
-                PathName = "Z 02 - Test Path"
-            });
-            _dbContext.Songs.Add(new Song
-            {
-                Title = "03 - Test Song",
-                Artist = "Test Artist",
-                Album = "Test Album",
-                Genre = "Test Genre - With a Very Long Name - And a Whole Bunch More Gibberish",
-                PathName = "Z 03 - Test Path"
-            });
+                Debug.WriteLine("\n=== ======================================== ===\n");
+                Debug.WriteLine("*** WARNING: This is a debug build.  The database will be deleted and recreated.  Change this later. (MauiProgram.cs)");
+                var _dbContext = new PlaylistContext();
+                _dbContext.Database.EnsureDeleted();
+                _dbContext.Database.EnsureCreated();
+                _dbContext.Playlists.Add(new Playlist { Name = "Default", Description = "The Default Play List" });
+                _dbContext.Playlists.Add(new Playlist { Name = "Test", Description = "The Test Play List" });
+                _dbContext.Playlists.Add(new Playlist { Name = "Test2", Description = "The Test2 Play List" });
+                _dbContext.Playlists.Add(new Playlist { Name = "Test3", Description = "The Test3 Play List" });
 
-            _dbContext.SaveChanges();
-            _dbContext.Dispose();
-            Debug.WriteLine("\n=== /Database Rebuild ====================== ===\n");
+                _dbContext.Songs.Add(new Song
+                {
+                    Title = "B - Test Song - With a Very Long Name",
+                    Artist = "Test Artist - With a Very Long Name",
+                    Album = "Test Album  - With a Very Long Name",
+                    Genre = "Test Genre - With a Very Long Name - And a Whole Bunch More Gibberish",
+                    PathName = "Z 02 - Test Path"
+                });
+                _dbContext.Songs.Add(new Song
+                {
+                    Title = "B - Test Song",
+                    Artist = "Test Artist",
+                    Album = "Test Album",
+                    Genre = "Test Genre - With a Very Long Name - And a Whole Bunch More Gibberish",
+                    PathName = "Z 03 - Test Path"
+                });
 
+                _dbContext.SaveChanges();
+                _dbContext.Dispose();
+                Debug.WriteLine("\n=== /Database Rebuild ====================== ===\n");
+            }
 
 
             return result;
